requirements.txt
pillow
requests
pymysql
aiomysql
asyncio
librosa
soundfile
SpeechRecognition
boto3
pytesseract
stripe
torch
opencv-python
pydub
beautifulsoup4
langdetect
nltk
transformers
openai
python-dotenv
python-telegram-bot
fer
tensorflow==2.17.0  # Add this line
timm
